test: use a different DataDirectoryValidator for tests
